+++ pkg/networkservice/metrics/stats/doc.go
@@ -1,4 +1,6 @@
-// Copyright (c) 2022 Cisco and/or its affiliates.
+// Copyright (c) 2021 Doc.ai and/or its affiliates.
+//
+// Copyright (c) 2024 Cisco and/or its affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -14,18 +16,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package stats provides chain elements for retrieving statistics from vpp interfaces
 package stats
-
-type statsOptions struct {
-	socket string
-}
-
-// Option is an option pattern for stats server/client
-type Option func(o *statsOptions)
-
-// WithSocket sets stats socket name
-func WithSocket(socket string) Option {
-	return func(o *statsOptions) {
-		o.socket = socket
-	}
-}
