+++ pkg/networkservice/metrics/stats/option.go
@@ -1,6 +1,4 @@
-// Copyright (c) 2021 Doc.ai and/or its affiliates.
-//
-// Copyright (c) 2024 Cisco and/or its affiliates.
+// Copyright (c) 2022-2024 Cisco and/or its affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -16,5 +14,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package metrics provides chain elements for retrieving statistics from vpp
-package metrics
+package stats
+
+type statsOptions struct {
+	socket string
+}
+
+// Option is an option pattern for stats server/client
+type Option func(o *statsOptions)
+
+// WithSocket sets stats socket name
+func WithSocket(socket string) Option {
+	return func(o *statsOptions) {
+		o.socket = socket
+	}
+}
